Replaced existing codebase with new project that builds as a library
Basically deleted the contents of this repository and replaced it with a
new library project from VS 2013. Contains the base class definitions
from the original code, but does not include the demo UI.

Also manually rearranged the classes defined in
ListViewDragDropManager.cs so that each class has its own file.